Moved Conway constrained specs to cardano-ledger-conway:testlib
