+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -116,6 +116,15 @@ library testlib
         Test.Cardano.Ledger.Conway.Binary.RoundTrip
         Test.Cardano.Ledger.Conway.Binary.Regression
         Test.Cardano.Ledger.Conway.Constrained
+        Test.Cardano.Ledger.Conway.Constrained.Instances
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Cert
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Deleg
+        Test.Cardano.Ledger.Conway.Constrained.Spec.GovCert
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Gov
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Ledger
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Pool
+        Test.Cardano.Ledger.Conway.Constrained.Spec.PParams
+        Test.Cardano.Ledger.Conway.Constrained.Spec.Utxo
         Test.Cardano.Ledger.Conway.Conformance.SpecTranslate
         Test.Cardano.Ledger.Conway.ImpTest
         Test.Cardano.Ledger.Conway.Imp
@@ -128,6 +137,7 @@ library testlib
         Test.Cardano.Ledger.Conway.Imp.RatifySpec
         Test.Cardano.Ledger.Conway.Proposals
         Test.Cardano.Ledger.Conway.TreeDiff
+
 
     visibility:       public
     hs-source-dirs:   testlib
@@ -143,8 +153,8 @@ library testlib
         bytestring,
         cardano-data:{cardano-data, testlib},
         containers,
-        constrained-generators,
         cryptonite,
+        constrained-generators,
         plutus-ledger-api,
         deepseq,
         microlens,
@@ -162,6 +172,7 @@ library testlib
         cardano-ledger-mary,
         cardano-ledger-shelley,
         cardano-strict-containers,
+        cardano-slotting:{cardano-slotting, testlib},
         data-default-class,
         generic-random,
         mtl,
