+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Constrained/Spec/Cert.hs
@@ -3,9 +3,8 @@
 
 -- | Specs necessary to generate, environment, state, and signal
 -- for the CERT rule
-module Test.Cardano.Ledger.Constrained.V2.Conway.CERT where
+module Test.Cardano.Ledger.Conway.Constrained.Spec.Cert where
 
-import Cardano.Ledger.Api
 import Cardano.Ledger.CertState
 import Cardano.Ledger.Conway.Rules
 import Cardano.Ledger.Conway.TxCert
@@ -13,11 +12,13 @@ import Cardano.Ledger.Shelley.API.Types
 
 import Constrained
 
-import Test.Cardano.Ledger.Constrained.V2.Conway
-import Test.Cardano.Ledger.Constrained.V2.Conway.DELEG
-import Test.Cardano.Ledger.Constrained.V2.Conway.GOVCERT
-import Test.Cardano.Ledger.Constrained.V2.Conway.POOL
-import Test.Cardano.Ledger.Constrained.V2.Conway.PParams
+import Test.Cardano.Ledger.Conway.Constrained.Instances
+import Test.Cardano.Ledger.Conway.Constrained.Spec.Deleg
+import Test.Cardano.Ledger.Conway.Constrained.Spec.GovCert
+import Test.Cardano.Ledger.Conway.Constrained.Spec.Pool
+import Test.Cardano.Ledger.Conway.Constrained.Spec.PParams
+import Cardano.Ledger.Conway (ConwayEra)
+import Cardano.Ledger.Crypto (StandardCrypto)
 
 certEnvSpec ::
   IsConwayUniv fn =>
