+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Constrained/Spec/Utxo.hs
@@ -3,13 +3,11 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE ImportQualifiedPost #-}
 {-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE TypeApplications #-}
 
 -- | Specs necessary to generate, environment, state, and signal
 -- for the UTXO rule
-module Test.Cardano.Ledger.Constrained.V2.Conway.UTXO where
+module Test.Cardano.Ledger.Conway.Constrained.Spec.Utxo where
 
-import Cardano.Ledger.Api
 import Cardano.Ledger.Babbage.TxOut
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Conway.PParams
@@ -25,8 +23,11 @@ import Lens.Micro
 
 import Constrained
 
-import Test.Cardano.Ledger.Constrained.V2.Conway
-import Test.Cardano.Ledger.Constrained.V2.Conway.PParams
+import Test.Cardano.Ledger.Conway.Constrained.Instances
+import Test.Cardano.Ledger.Conway.Constrained.Spec.PParams
+import Cardano.Ledger.Conway (ConwayEra)
+import Cardano.Ledger.Crypto (StandardCrypto)
+import Cardano.Ledger.Conway.Core (EraTx(..), ppMaxCollateralInputsL)
 
 utxoEnvSpec :: IsConwayUniv fn => Spec fn (UtxoEnv (ConwayEra StandardCrypto))
 utxoEnvSpec =
